+++ src/song/mod.rs
@@ -17,7 +17,7 @@ use crate::cue::CueInfo;
 use crate::misc::LoudnessDesc;
 use crate::temporal::BPMDesc;
 use crate::timbral::{SpectralDesc, ZeroCrossingRateDesc};
-use crate::{BlissError, BlissResult, FeaturesVersion, FEATURES_VERSION, SAMPLE_RATE};
+use crate::{BlissError, BlissResult, FeaturesVersion, SAMPLE_RATE};
 use core::ops::Index;
 use ndarray::{arr1, Array1};
 use std::convert::TryInto;
@@ -138,6 +138,10 @@ pub const NUMBER_FEATURES: usize = AnalysisIndex::COUNT;
 /// directly in this code.
 pub struct Analysis {
     pub(crate) internal_analysis: [f32; NUMBER_FEATURES],
+    // Version of the features the song was analyzed with.
+    /// It is bumped every time a change is introduced in the
+    /// features that makes them incompatible with previous versions.
+    pub features_version: FeaturesVersion,
 }
 
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
@@ -146,7 +150,7 @@ pub struct Analysis {
 /// of a song.
 pub struct AnalysisOptions {
     /// The version of the features that should be used for analysis.
-    /// Should be kept as the default [FEATURES_VERSION](bliss_audio::FEATURES_VERSION).
+    /// Should be kept as the default [FeaturesVersion::LATEST](bliss_audio::FeaturesVersion::LATEST).
     pub features_version: FeaturesVersion,
     /// The number of computer cores that should be used when performing the
     /// analysis of multiple songs.
@@ -157,7 +161,7 @@ impl Default for AnalysisOptions {
     fn default() -> Self {
         let cores = thread::available_parallelism().unwrap_or(NonZeroUsize::new(1).unwrap());
         AnalysisOptions {
-            features_version: FEATURES_VERSION,
+            features_version: FeaturesVersion::LATEST,
             number_cores: cores,
         }
     }
@@ -173,7 +177,10 @@ impl Index<AnalysisIndex> for Analysis {
 
 impl fmt::Debug for Analysis {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut debug_struct = f.debug_struct("Analysis");
+        let mut debug_struct = f.debug_struct(&format!(
+            "Analysis (Version {})",
+            self.features_version as u16
+        ));
         for feature in AnalysisIndex::iter() {
             debug_struct.field(&format!("{feature:?}"), &self[feature]);
         }
@@ -188,9 +195,10 @@ impl Analysis {
     /// Usually not needed, unless you have already computed and stored
     /// features somewhere, and need to recreate a Song with an already
     /// existing Analysis yourself.
-    pub fn new(analysis: [f32; NUMBER_FEATURES]) -> Analysis {
+    pub fn new(analysis: [f32; NUMBER_FEATURES], features_version: FeaturesVersion) -> Analysis {
         Analysis {
             internal_analysis: analysis,
+            features_version,
         }
     }
 
@@ -240,93 +248,7 @@ impl Song {
      * you the raw sample array as it should look like, if you're not into computing checksums)
      **/
     pub fn analyze(sample_array: &[f32]) -> BlissResult<Analysis> {
-        let largest_window = vec![
-            BPMDesc::WINDOW_SIZE,
-            ChromaDesc::WINDOW_SIZE,
-            SpectralDesc::WINDOW_SIZE,
-            LoudnessDesc::WINDOW_SIZE,
-        ]
-        .into_iter()
-        .max()
-        .unwrap();
-        if sample_array.len() < largest_window {
-            return Err(BlissError::AnalysisError(String::from(
-                "empty or too short song.",
-            )));
-        }
-
-        thread::scope(|s| -> BlissResult<Analysis> {
-            let child_tempo = s.spawn(|| -> BlissResult<f32> {
-                let mut tempo_desc = BPMDesc::new(SAMPLE_RATE)?;
-                let windows = sample_array
-                    .windows(BPMDesc::WINDOW_SIZE)
-                    .step_by(BPMDesc::HOP_SIZE);
-
-                for window in windows {
-                    tempo_desc.do_(window)?;
-                }
-                Ok(tempo_desc.get_value())
-            });
-
-            let child_chroma = s.spawn(|| -> BlissResult<Vec<f32>> {
-                let mut chroma_desc = ChromaDesc::new(SAMPLE_RATE, 12);
-                chroma_desc.do_(sample_array)?;
-                Ok(chroma_desc.get_values())
-            });
-
-            #[allow(clippy::type_complexity)]
-            let child_timbral = s.spawn(|| -> BlissResult<(Vec<f32>, Vec<f32>, Vec<f32>)> {
-                let mut spectral_desc = SpectralDesc::new(SAMPLE_RATE)?;
-                let windows = sample_array
-                    .windows(SpectralDesc::WINDOW_SIZE)
-                    .step_by(SpectralDesc::HOP_SIZE);
-                for window in windows {
-                    spectral_desc.do_(window)?;
-                }
-                let centroid = spectral_desc.get_centroid();
-                let rolloff = spectral_desc.get_rolloff();
-                let flatness = spectral_desc.get_flatness();
-                Ok((centroid, rolloff, flatness))
-            });
-
-            let child_zcr = s.spawn(|| -> BlissResult<f32> {
-                let mut zcr_desc = ZeroCrossingRateDesc::default();
-                zcr_desc.do_(sample_array);
-                Ok(zcr_desc.get_value())
-            });
-
-            let child_loudness = s.spawn(|| -> BlissResult<Vec<f32>> {
-                let mut loudness_desc = LoudnessDesc::default();
-                let windows = sample_array.chunks(LoudnessDesc::WINDOW_SIZE);
-
-                for window in windows {
-                    loudness_desc.do_(window);
-                }
-                Ok(loudness_desc.get_value())
-            });
-
-            // Non-streaming approach for that one
-            let tempo = child_tempo.join().unwrap()?;
-            let chroma = child_chroma.join().unwrap()?;
-            let (centroid, rolloff, flatness) = child_timbral.join().unwrap()?;
-            let loudness = child_loudness.join().unwrap()?;
-            let zcr = child_zcr.join().unwrap()?;
-
-            let mut result = vec![tempo, zcr];
-            result.extend_from_slice(&centroid);
-            result.extend_from_slice(&rolloff);
-            result.extend_from_slice(&flatness);
-            result.extend_from_slice(&loudness);
-            result.extend_from_slice(&chroma);
-            let array: [f32; NUMBER_FEATURES] = result.try_into().map_err(|_| {
-                BlissError::AnalysisError(
-                    "Too many or too little features were provided at the end of
-                        the analysis."
-                        .to_string(),
-                )
-            })?;
-            Ok(Analysis::new(array))
-        })
+        Self::analyze_with_options(sample_array, &AnalysisOptions::default())
     }
 
     /**
@@ -427,7 +349,7 @@ impl Song {
                         .to_string(),
                 )
             })?;
-            Ok(Analysis::new(array))
+            Ok(Analysis::new(array, analysis_options.features_version))
         })
     }
 }
@@ -440,7 +362,7 @@ mod tests {
     #[cfg(feature = "ffmpeg")]
     use crate::decoder::Decoder as DecoderTrait;
     #[cfg(feature = "ffmpeg")]
-    use crate::FEATURES_VERSION;
+    use crate::FeaturesVersion;
     use pretty_assertions::assert_eq;
     #[cfg(feature = "ffmpeg")]
     use std::path::Path;
@@ -494,7 +416,7 @@ mod tests {
         for (x, y) in song.analysis.as_vec().iter().zip(expected_analysis) {
             assert!(0.01 > (x - y).abs());
         }
-        assert_eq!(FEATURES_VERSION, song.features_version);
+        assert_eq!(FeaturesVersion::LATEST, song.features_version);
     }
 
     #[test]
@@ -550,7 +472,7 @@ mod tests {
         for (x, y) in song.analysis.as_vec().iter().zip(expected_analysis) {
             assert!(0.01 > (x - y).abs());
         }
-        assert_eq!(FEATURES_VERSION, song.features_version);
+        assert_eq!(FeaturesVersion::LATEST, song.features_version);
     }
 
     #[test]
@@ -566,7 +488,7 @@ mod tests {
     fn test_debug_analysis() {
         let song = Decoder::song_from_path("data/s16_mono_22_5kHz.flac").unwrap();
         assert_eq!(
-            "Analysis { Tempo: 0.3846389, Zcr: -0.849141, MeanSpectralCentroid: -0.75481045, StdDeviationSpectralCentroid: -0.8790748, MeanSpectralRolloff: -0.63258266, StdDeviationSpectralRolloff: -0.7258959, MeanSpectralFlatness: -0.7757379, StdDeviationSpectralFlatness: -0.8146726, MeanLoudness: 0.2716726, StdDeviationLoudness: 0.25779057, Chroma1: -0.34292513, Chroma2: -0.62803423, Chroma3: -0.28095096, Chroma4: 0.08686459, Chroma5: 0.24446082, Chroma6: -0.5723257, Chroma7: 0.23292065, Chroma8: 0.19981146, Chroma9: -0.58594406, Chroma10: -0.06784296 } /* [0.3846389, -0.849141, -0.75481045, -0.8790748, -0.63258266, -0.7258959, -0.7757379, -0.8146726, 0.2716726, 0.25779057, -0.34292513, -0.62803423, -0.28095096, 0.08686459, 0.24446082, -0.5723257, 0.23292065, 0.19981146, -0.58594406, -0.06784296] */",
+            "Analysis (Version 2) { Tempo: 0.3846389, Zcr: -0.849141, MeanSpectralCentroid: -0.75481045, StdDeviationSpectralCentroid: -0.8790748, MeanSpectralRolloff: -0.63258266, StdDeviationSpectralRolloff: -0.7258959, MeanSpectralFlatness: -0.7757379, StdDeviationSpectralFlatness: -0.8146726, MeanLoudness: 0.2716726, StdDeviationLoudness: 0.25779057, Chroma1: -0.34292513, Chroma2: -0.62803423, Chroma3: -0.28095096, Chroma4: 0.08686459, Chroma5: 0.24446082, Chroma6: -0.5723257, Chroma7: 0.23292065, Chroma8: 0.19981146, Chroma9: -0.58594406, Chroma10: -0.06784296 } /* [0.3846389, -0.849141, -0.75481045, -0.8790748, -0.63258266, -0.7258959, -0.7757379, -0.8146726, 0.2716726, 0.25779057, -0.34292513, -0.62803423, -0.28095096, 0.08686459, 0.24446082, -0.5723257, 0.23292065, 0.19981146, -0.58594406, -0.06784296] */",
             format!("{:?}", song.analysis),
         );
     }